# apps/inngest-app/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Inngest App API",
    "version": "1.0.0",
    "description": "API documentation for the Inngest app"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your bearer token in the format: Bearer {token}"
      }
    },
...
  },
  "paths": {
...
    "/api/marketplace/workflows": {
      "get": {
        "summary": "Browse available workflow templates",
        "description": "Returns paginated list of workflow templates available for installation from the marketplace.",
        "tags": [
          "Marketplace"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            },
            "description": "Number of templates per page"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search templates by name or description"
          },
          {
            "in": "query",
            "name": "provider",
            "schema": {
              "type": "string",
              "enum": [
                "GOOGLE",
                "SLACK",
                "HUBSPOT",
                "FIRECRAWL",
                "CUSTOM"
              ]
            },
            "description": "Filter by required provider"
          },
          {
            "in": "query",
            "name": "pricing",
            "schema": {
              "type": "string",
              "enum": [
                "free",
                "paid"
              ]
            },
            "description": "Filter by pricing type"
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            },
            "description": "Filter by category"
          },
          {
            "in": "query",
            "name": "featured",
            "schema": {
              "type": "boolean"
            },
            "description": "Show only featured templates"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of workflow templates",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              },
                              "description": {
                                "type": "string"
                              },
                              "canBeScheduled": {
                                "type": "boolean"
                              },
                              "requiredProviders": {
                                "type": "array",
                                "items": {
                                  "type": "string"
                                }
                              },
                              "requiredScopes": {
                                "type": "object",
                                "additionalProperties": {
                                  "type": "array",
                                  "items": {
                                    "type": "string"
                                  }
                                },
                                "description": "Required OAuth scopes per provider"
                              },
                              "type": {
                                "type": "string",
                                "enum": [
                                  "marketplace",
                                  "custom"
                                ]
                              },
                              "pricing": {
                                "type": "object",
                                "properties": {
                                  "type": {
                                    "type": "string",
                                    "enum": [
                                      "free",
                                      "paid"
                                    ]
                                  },
                                  "price": {
                                    "type": "number",
                                    "description": "Price in cents"
                                  },
                                  "currency": {
                                    "type": "string",
                                    "enum": [
                                      "USD",
                                      "EUR",
                                      "GBP"
                                    ]
                                  },
                                  "billingPeriod": {
                                    "type": "string",
                                    "enum": [
                                      "one_time",
                                      "monthly",
                                      "yearly"
                                    ]
                                  }
                                }
                              },
                              "category": {
                                "type": "string"
                              },
                              "tags": {
                                "type": "array",
                                "items": {
                                  "type": "string"
                                }
                              },
                              "featured": {
                                "type": "boolean"
                              },
                              "author": {
                                "type": "string"
                              },
                              "version": {
                                "type": "string"
                              },
                              "isInstalled": {
                                "type": "boolean"
                              },
                              "canInstall": {
                                "type": "boolean"
                              }
                            }
                          }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "page": {
                              "type": "integer"
                            },
                            "limit": {
                              "type": "integer"
                            },
                            "total": {
                              "type": "integer"
                            },
                            "totalPages": {
                              "type": "integer"
                            },
                            "hasNext": {
                              "type": "boolean"
                            },
                            "hasPrev": {
                              "type": "boolean"
                            }
                          }
                        }
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/realtime/subscription-token": {
      "post": {
        "summary": "Get realtime subscription token for a workflow",
        "description": "Returns a short-lived token to subscribe to realtime updates for a specific workflow's channel.",
        "tags": [
          "Realtime"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "workflowId": {
                    "type": "string",
                    "description": "The workflow ID to subscribe to"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subscription token generated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing workflowId"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User or workflow not found"
          },
          "500": {
            "description": "Failed to generate subscription token"
          }
        }
      }
    },
...
  },
...
}